pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Chatbot-Using-RAG"
version = "0.1.0"
description = "A brief description of your project"
authors = [{ name = "Divyansh Vasal", email = "divyanshvasal2511@gmail.com" }]
dependencies = [
    "_libgcc_mutex=0.1=main",
    "_openmp_mutex=5.1=1_gnu",
    "asttokens=2.0.5=pyhd3eb1b0_0",
    "bzip2=1.0.8=h5eee18b_6",
    "ca-certificates=2024.11.26=h06a4308_0",
    "comm=0.2.1=py311h06a4308_0",
    "debugpy=1.6.7=py311h6a678d5_0",
    "decorator=5.1.1=pyhd3eb1b0_0",
    "executing=0.8.3=pyhd3eb1b0_0",
    "greenlet=3.1.1=py311h6a678d5_0",
    "ipykernel=6.29.5=py311h06a4308_0",
    "ipython=8.30.0=py311h06a4308_0",
    "jedi=0.19.2=py311h06a4308_0",
    "jupyter_client=8.6.0=py311h06a4308_0",
    "jupyter_core=5.7.2=py311h06a4308_0",
    "ld_impl_linux-64=2.40=h12ee557_0",
    "libffi=3.4.4=h6a678d5_1",
    "libgcc-ng=11.2.0=h1234567_1",
    "libgomp=11.2.0=h1234567_1",
    "libsodium=1.0.18=h7b6447c_0",
    "libstdcxx-ng=11.2.0=h1234567_1",
    "libuuid=1.41.5=h5eee18b_0",
    "matplotlib-inline=0.1.6=py311h06a4308_0",
    "ncurses=6.4=h6a678d5_0",
    "nest-asyncio=1.6.0=py311h06a4308_0",
    "openssl=3.0.15=h5eee18b_0",
    "packaging=24.2=py311h06a4308_0",
    "parso=0.8.4=py311h06a4308_0",
    "pexpect=4.8.0=pyhd3eb1b0_3",
    "platformdirs=3.10.0=py311h06a4308_0",
    "prompt-toolkit=3.0.43=py311h06a4308_0",
    "prompt_toolkit=3.0.43=hd3eb1b0_0",
    "psutil=5.9.0=py311h5eee18b_0",
    "ptyprocess=0.7.0=pyhd3eb1b0_2",
    "pure_eval=0.2.2=pyhd3eb1b0_0",
    "pygments=2.15.1=py311h06a4308_1",
    "python=3.11.11=he870216_0",
    "python-dateutil=2.9.0post0=py311h06a4308_2",
    "pyzmq=26.2.0=py311h6a678d5_0",
    "readline=8.2=h5eee18b_0",
    "setuptools=75.1.0=py311h06a4308_0",
    "six=1.16.0=pyhd3eb1b0_1",
    "sqlalchemy=2.0.34=py311h00e1ef3_0",
    "sqlite=3.45.3=h5eee18b_0",
    "stack_data=0.2.0=pyhd3eb1b0_0",
    "tk=8.6.14=h39e8969_0",
    "tornado=6.4.2=py311h5eee18b_0",
    "traitlets=5.14.3=py311h06a4308_0",
    "typing-extensions=4.12.2=py311h06a4308_0",
    "typing_extensions=4.12.2=py311h06a4308_0",
    "wcwidth=0.2.5=pyhd3eb1b0_0",
    "wheel=0.44.0=py311h06a4308_0",
    "xz=5.4.6=h5eee18b_1",
    "zeromq=4.3.5=h6a678d5_0",
    "zlib=1.2.13=h5eee18b_1",
    "absl-py==2.1.0",
    "accelerate==1.2.1",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "astunparse==1.6.3",
    "attrs==24.3.0",
    "beautifulsoup4==4.12.3",
    "blis==1.1.0",
    "cachetools==5.5.0",
    "catalogue==2.0.10",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.0",
    "click==8.1.8",
    "cloudpathlib==0.20.0",
    "confection==0.1.5",
    "cymem==2.0.10",
    "datasets==3.2.0",
    "dill==0.3.8",
    "en-core-web-sm==3.8.0",
    "evaluate==0.4.3",
    "fastapi==0.115.6",
    "filelock==3.13.1",
    "flatbuffers==24.3.25",
    "frozenlist==1.5.0",
    "fsspec==2024.6.1",
    "gast==0.6.0",
    "google-auth==2.37.0",
    "google-auth-oauthlib==1.2.1",
    "google-pasta==0.2.0",
    "grpcio==1.68.1",
    "h11==0.14.0",
    "h5py==3.12.1",
    "huggingface-hub==0.27.0",
    "idna==3.10",
    "intel-cmplr-lib-rt==2025.0.2",
    "intel-cmplr-lib-ur==2025.0.2",
    "intel-cmplr-lic-rt==2025.0.2",
    "intel-pti==0.10.0",
    "intel-sycl-rt==2025.0.2",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "keras==2.15.0",
    "langcodes==3.5.0",
    "language-data==1.3.0",
    "libclang==18.1.1",
    "marisa-trie==1.2.1",
    "markdown==3.7",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "mdurl==0.1.2",
    "ml-dtypes==0.2.0",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "murmurhash==1.0.11",
    "networkx==3.3",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "opt-einsum==3.4.0",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "pip==24.3.1",
    "preshed==3.0.9",
    "propcache==0.2.1",
    "protobuf==4.25.5",
    "psycopg2-binary==2.9.10",
    "pyarrow==18.1.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "python-dotenv==1.0.1",
    "pytorch-triton-xpu==3.2.0",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==13.9.4",
    "rouge-score==0.1.2",
    "rsa==4.9",
    "safetensors==0.4.5",
    "scikit-learn==1.6.0",
    "scipy==1.14.1",
    "shellingham==1.5.4",
    "smart-open==7.1.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "spacy==3.8.3",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.5.0",
    "starlette==0.41.3",
    "sympy==1.13.1",
    "tcmlib==1.2.0",
    "tensorboard==2.15.2",
    "tensorboard-data-server==0.7.2",
    "tensorboardx==2.6.2.2",
    "tensorflow==2.15.0",
    "tensorflow-estimator==2.15.0",
    "tensorflow-io-gcs-filesystem==0.37.1",
    "termcolor==2.5.0",
    "thinc==8.3.3",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "torch==2.6.0+xpu",
    "torchaudio==2.6.0+xpu",
    "torchvision==0.21.0+xpu",
    "tqdm==4.67.1",
    "transformers==4.47.1",
    "triton==3.2.0",
    "typer==0.15.1",
    "tzdata==2024.2",
    "umf==0.9.1",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "wasabi==1.1.3",
    "weasel==0.4.1",
    "werkzeug==3.1.3",
    "wget==3.2",
    "wrapt==1.14.1",
    "xxhash==3.5.0",
    "yarl==1.18.3",
]
